Testing, Refactoring and Clean-Up
Write additional tests for checking restricted access to views depending on which group a user belongs to.
Refactor the view tests to align with the new user model as well as the newly added user permissions
Add a custom UserManager for creating users and super-users manually without needing a username. Overrides the methods of the built-in Django library AbstractUser